Move the unit combinators to Unit.Algebra.
